// src/2-sales/domain/enums/status-pedido.enum.ts
export enum StatusPedido {
  PAGAMENTO_PENDENTE = 'PAGAMENTO_PENDENTE',
  PAGO = 'PAGO',
  PENDENTE = 'PENDENTE',
  EM_PREPARO = 'EM_PREPARO',
  ENVIADO = 'ENVIADO',
  ENTREGUE = 'ENTREGUE',
  CANCELADO = 'CANCELADO',
  REEMBOLSO_SOLICITADO = 'REEMBOLSO_SOLICITADO',
  REEMBOLSADO = 'REEMBOLSADO',
}